django_api_accretion/django_api_accretion/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
boto3==1.34.113; python_version >= '3.8'
botocore==1.34.113; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cryptography==42.0.7; python_version >= '3.7'
defusedxml==0.8.0rc2; python_version >= '3.6'
django==5.0.6; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-debug-toolbar==4.4.1; python_version >= '3.8'
django-ses==4.1.0; python_version >= '3.8' and python_version < '4.0'
django-storages==1.14.3; python_version >= '3.7'
django-templated-mail==1.1.1
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djoser==2.2.2; python_version >= '3.8' and python_version < '4.0'
gunicorn==22.0.0; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.0; python_version >= '3.7'
pillow==10.3.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python3-openid==3.2.0
pytz==2024.1
requests==2.32.2; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
s3transfer==0.10.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.4.1; python_version >= '3.8'
social-auth-core==4.5.4; python_version >= '3.8'
soupsieve==2.5; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.12.0; python_version < '3.11'
urllib3==2.2.1; python_version >= '3.10'
